Fixing config, creating dpapi encryption
Adding the credentials path to the back-end config, to make it easier to install the website (I would have to modify the path many files, this is better for customization), same for users_win.

Improving 'root_credentials', 'secret_manager', creating new functions for encryption with dpapi in the helpers (again to make installation easier, because for website to work correctly we have to encrypt the credentials with dpapi).